% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HiContactsData.R
\name{HiContactsData}
\alias{HiContactsData}
\title{HiContactsData}
\usage{
HiContactsData(sample = NULL, format = NULL)
}
\arguments{
\item{sample}{sample}

\item{format}{format}
}
\value{
Local path of the queried file cached with BiocFileCache.
}
\description{
Downloads different types of Hi-C processed files
(cool, mcool, pairs.gz) and returns the path of the cached file.
}
\examples{
HiContactsData(sample = 'yeast_wt', format = 'cool')
}
