% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HiContactsData.R
\name{HiContactsDataFiles}
\alias{HiContactsDataFiles}
\title{HiContactsDataFiles}
\usage{
HiContactsDataFiles()
}
\value{
data.frame
}
\description{
List of all available files provided by HiContactsData.
}
\examples{
HiContactsDataFiles()
}
